[Refactor, Chore] 디자인 시스템 내부 파일 이동 및 수정, 모듈명 변경 (#17)
* [chore] : FunchCaption ui 패키지로 이동

* [chore] : FunchIconButton 컴포넌트 Button.kt 파일로 이동

* [refactor] : Preview 가시성 지시자 부여 및 Preview 이름 변경

* [refactor] : FunchErrorCaption UI 수정

* [chore] : home 모듈명 profile로 변경

* [feat] : @Immutable 어노테이션 추가 - 리뷰 반영
