+++ core/designsystem/src/main/java/com/moya/funch/component/Button.kt
@@ -1,6 +1,11 @@
 package com.moya.funch.component
 
+import androidx.annotation.DrawableRes
+import androidx.compose.foundation.Indication
+import androidx.compose.foundation.LocalIndication
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -9,24 +14,31 @@ import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Immutable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import com.moya.funch.icon.FunchIconAsset
 import com.moya.funch.modifier.clickableSingle
 import com.moya.funch.modifier.neonSign
 import com.moya.funch.theme.FunchTheme
 import com.moya.funch.theme.Gray400
+import com.moya.funch.theme.Gray500
 import com.moya.funch.theme.Gray800
 import com.moya.funch.theme.Gray900
 import com.moya.funch.theme.Lemon500
@@ -41,6 +53,13 @@ enum class FunchButtonType(val shape: Shape, val contentVerticalPadding: Dp, val
     Small(RoundedCornerShape(12.dp), 12.dp, funchTypography.b),
     XSmall(RoundedCornerShape(12.dp), 8.dp, funchTypography.b);
 }
+
+@Immutable
+data class FunchIcon(
+    @DrawableRes val resId: Int,
+    val description: String,
+    val tint: Color,
+)
 
 @Composable
 fun FunchMainButton(
@@ -146,6 +165,40 @@ fun FunchSubButton(
         }
     }
 }
+
+@Composable
+fun FunchIconButton(
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit,
+    funchIcon: FunchIcon,
+    backgroundColor: Color = Color.Transparent,
+    roundedCornerShape: RoundedCornerShape = RoundedCornerShape(0.dp),
+    indication: Indication? = LocalIndication.current,
+    interactionSource: MutableInteractionSource = remember { MutableInteractionSource() },
+) {
+    Box(
+        modifier = modifier
+            .background(
+                color = backgroundColor,
+                shape = roundedCornerShape,
+            )
+            .clip(roundedCornerShape)
+            .clickable(
+                onClick = onClick,
+                indication = indication,
+                interactionSource = interactionSource,
+            ),
+        contentAlignment = Alignment.Center,
+    ) {
+        Icon(
+            painter = painterResource(id = funchIcon.resId),
+            contentDescription = funchIcon.description,
+            tint = funchIcon.tint,
+        )
+    }
+}
+
+/*============================== Preview =================================*/
 
 @Preview(name = "main button size", showBackground = true, widthDp = 360)
 @Composable
@@ -307,4 +360,47 @@ private fun Preview4() {
             )
         }
     }
+}
+
+@Preview("Icon Large Button", showBackground = true, backgroundColor = 0xFFFFFF)
+@Composable
+private fun Preview5() {
+    FunchIconButton(
+        modifier = Modifier.size(40.dp),
+        roundedCornerShape = RoundedCornerShape(12.dp),
+        backgroundColor = Gray500,
+        onClick = { /*TODO*/ },
+        funchIcon = FunchIcon(
+            resId = FunchIconAsset.Search.search_24,
+            description = "",
+            tint = Yellow500,
+        ),
+    )
+}
+
+@Preview("Icon Medium Button", showBackground = true)
+@Composable
+private fun Preview6() {
+    FunchIconButton(
+        onClick = { /*TODO*/ },
+        funchIcon = FunchIcon(
+            resId = FunchIconAsset.Search.search_24,
+            description = "",
+            tint = Gray400,
+        ),
+    )
+}
+
+@Preview("Icon Small Button", showBackground = true)
+@Composable
+private fun Preview7() {
+    FunchIconButton(
+        onClick = { /*TODO*/ },
+        funchIcon = FunchIcon(
+            resId = FunchIconAsset.Search.search_16,
+            description = "",
+            tint = Gray400,
+        ),
+        indication = null,
+    )
 }
