--- feature/home/src/main/java/com/moya/funch/HomeScreen.kt
@@ -1,11 +1,0 @@
-package com.moya.funch
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun HomeScreen() {
-    Text(
-        text = "Hello HomeScreen!",
-    )
-}
