+++ feature/profile/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.moya.funch.home">
+<manifest package="com.moya.funch.profile">
 
 </manifest>
