+++ feature/profile/src/main/java/com/moya/funch/MyProfileScreen.kt
@@ -1,0 +1,8 @@
+package com.moya.funch
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun MyProfileScreen() {
+    // @GunHyung TODO : 내 정보 화면 구현
+}
