+++ settings.gradle.kts
@@ -24,6 +24,5 @@ include(":core:designsystem")
 //include(":core:network")
 
 // feature
-include(":feature:home")
-// include(":feature:profile")
+include(":feature:profile")
 // include(":feature:match")
